test: seperate cloud-specific tests from other e2e code
